+++ packages/solid/src/components/listbox/listbox-root-provider.tsx
@@ -1,10 +1,11 @@
 import { mergeProps } from '@zag-js/solid'
 import type { JSX } from 'solid-js/jsx-runtime'
 import { createSplitProps } from '../../utils/create-split-props'
-import type { Assign, CollectionItem } from '../../types'
+import type { Assign } from '../../types'
 import { type HTMLProps, type PolymorphicProps, ark } from '../factory'
 import type { UseListboxReturn } from './use-listbox'
 import { ListboxProvider } from './use-listbox-context'
+import type { CollectionItem } from '../collection'
 
 interface RootProviderProps<T extends CollectionItem> {
   value: UseListboxReturn<T>