+++ packages/solid/src/components/select/select-root-provider.tsx
@@ -1,11 +1,12 @@
 import { mergeProps } from '@zag-js/solid'
 import type { JSX } from 'solid-js/jsx-runtime'
-import type { Assign, CollectionItem } from '../../types'
+import type { Assign } from '../../types'
 import { createSplitProps } from '../../utils/create-split-props'
 import { type HTMLProps, type PolymorphicProps, ark } from '../factory'
 import { PresenceProvider, type UsePresenceProps, splitPresenceProps, usePresence } from '../presence'
 import type { UseSelectReturn } from './use-select'
 import { SelectProvider } from './use-select-context'
+import type { CollectionItem } from '../collection'
 
 interface RootProviderProps<T extends CollectionItem> {
   value: UseSelectReturn<T>