+++ packages/svelte/src/lib/types.ts
@@ -3,9 +3,6 @@ import type { HTMLAttributes, SvelteHTMLElements } from 'svelte/elements'
 
 export type Assign<T, U> = Omit<T, keyof U> & U
 export type Optional<T, K extends keyof T> = Pick<Partial<T>, K> & Omit<T, K>
-
-export type CollectionItem = string | object
-
 export type Accessor<T> = () => T
 
 export type HTMLTag = keyof SvelteHTMLElements